Moving BaseSpringUnitTest to root.
