+++ core/src/test/java/org/sanelib/eboss/ApplicationContextTest.java
